Move los_sync_query tests into a separate file and add tests for the change
